requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt pyproject.toml
#
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   openai
beautifulsoup4==4.12.0
    # via telegram-bot-app (pyproject.toml)
blinker==1.9.0
    # via flask
certifi==2025.7.9
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via flask
colorama==0.4.6
    # via
    #   click
    #   tqdm
distro==1.9.0
    # via openai
exceptiongroup==1.3.0
    # via anyio
flask==2.3.2
    # via telegram-bot-app (pyproject.toml)
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.25.2
    # via
    #   openai
    #   python-telegram-bot
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
jiter==0.10.0
    # via openai
markupsafe==3.0.2
    # via
    #   jinja2
    #   werkzeug
openai==1.95.1
    # via telegram-bot-app (pyproject.toml)
pydantic==2.11.7
    # via openai
pydantic-core==2.33.2
    # via pydantic
python-dotenv==1.0.0
    # via telegram-bot-app (pyproject.toml)
python-telegram-bot[webhooks]==20.7
    # via telegram-bot-app (pyproject.toml)
requests==2.31.0
    # via telegram-bot-app (pyproject.toml)
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
soupsieve==2.7
    # via beautifulsoup4
tornado==6.3.3
    # via python-telegram-bot
tqdm==4.67.1
    # via openai
typing-extensions==4.14.1
    # via
    #   anyio
    #   exceptiongroup
    #   openai
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
urllib3==2.5.0
    # via requests
werkzeug==3.1.3
    # via flask
